+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+# v1.7.0
+
+## Enhancements
+
 # v1.6.0
 
 ## Enhancements